Move the spreadable route-table factories into route-tables.ts
entityTabRoutes had pushed entity-pages.ts past the ~400-line target.
Both factories (and the CrudHandlers contract) now live in one focused
module, so entity-pages.ts is back to 382 lines and sections import
their route spreads from a single place.
