+++ src/features/admin/owner-crud.ts
@@ -3,6 +3,7 @@ import {
   type FormGuard,
 } from "#routes/admin/confirmation.ts";
 import type { EditErrorRenderer } from "#routes/admin/entity-write-tab.ts";
+import type { CrudHandlers } from "#routes/admin/route-tables.ts";
 import {
   AUTH_FORM,
   type AuthSession,
@@ -18,11 +19,9 @@ import {
 /* jscpd:ignore-start */
 import { type IdRouteHandler, idRouteFor } from "#routes/entity.ts";
 import { errorRedirect, notFoundResponse, redirect } from "#routes/response.ts";
-import type { TypedRouteHandler } from "#routes/router.ts";
 /* jscpd:ignore-end */
 import { logActivity } from "#shared/db/activity-log.ts";
 import { getFlash } from "#shared/flash-context.ts";
-import type { RequestRoute } from "#shared/response-steps.ts";
 import type {
   DeleteResult,
   NamedOperations,
@@ -34,44 +33,6 @@ type OperationFailure = Exclude<
   DeleteResult | UpdateResult<unknown>,
   { ok: true }
 >;
-
-/** The handler bundle a CRUD factory returns, bindable via
- * {@link crudRoutes} or one key at a time. */
-export interface CrudHandlers {
-  createPost: RequestRoute;
-  deleteGet: IdRouteHandler;
-  deletePost: IdRouteHandler;
-  editPost: IdRouteHandler;
-  listGet: RequestRoute;
-  newGet: RequestRoute;
-}
-
-/** The six routes a standard CRUD section binds. */
-export type CrudRoutes<Base extends string> = {
-  [K in
-    | `GET ${Base}`
-    | `GET ${Base}/new`
-    | `GET ${Base}/:id/delete`
-    | `POST ${Base}`
-    | `POST ${Base}/:id/delete`
-    | `POST ${Base}/:id/edit`]: TypedRouteHandler<K>;
-};
-
-/** Bind a CRUD handler bundle under its section's six standard routes, ready
- * to spread into the route table. A section with a bespoke step spreads this
- * first and restates just that key (e.g. a custom `:id/edit` POST). */
-export const crudRoutes = <Base extends string>(
-  base: Base,
-  crud: CrudHandlers,
-): CrudRoutes<Base> =>
-  ({
-    [`GET ${base}`]: crud.listGet,
-    [`GET ${base}/new`]: crud.newGet,
-    [`GET ${base}/:id/delete`]: crud.deleteGet,
-    [`POST ${base}`]: crud.createPost,
-    [`POST ${base}/:id/delete`]: crud.deletePost,
-    [`POST ${base}/:id/edit`]: crud.editPost,
-  }) as CrudRoutes<Base>;
 
 /** Resolve one CRUD operation through its success or failure response path. */
 export const operationResponse = async <Success extends { ok: true }, Output>(
